Move to a newer version of jgit
This is to get a newer version of jsch (a transitive dependency of
jgit).  Older versions of jsch don't appear to support the key exchange
algorithms required by github.  This means that when this plugin tries
to push tags to a github repo it will fail with the error `Algorithm
negotiation fail`.